wlwodlf/requirements.txt
streamlit
google-generativeai
python-dotenv
requests
beautifulsoup4
youtube-transcript-api
matplotlib
